leetcode/008backtracking/0332. Reconstruct Itinerary.go
package leetcode

import "sort"

// 给定一个机票的字符串二维数组 [from, to]，子数组中的两个成员分别表示飞机出发和降落的机场地点，对该行程进行重新规划排序。
// 所有这些机票都属于一个从 JFK（肯尼迪国际机场）出发的先生，所以该行程必须从 JFK 开始。

type pair struct {
	target  string
	visited bool
}
type pairs []*pair

func (p pairs) Len() int {
	return len(p)
}
func (p pairs) Swap(i, j int) {
	p[i], p[j] = p[j], p[i]
}

func (p pairs) Less(i, j int) bool {
	return p[i].target < p[j].target
}

func findItinerary(tickets [][]string) []string {
	result := []string{}
	// map[出发机场] pair{目的地,是否被访问过}
	targets := make(map[string]pairs)
	for _, ticket := range tickets {
		if targets[ticket[0]] == nil {
			targets[ticket[0]] = make(pairs, 0)
		}
		// 相同出发机场的机票加入pairs
		targets[ticket[0]] = append(targets[ticket[0]], &pair{target: ticket[1], visited: false})
	}
	// 对相同出发机场的机票进行排序
	for k := range targets {
		sort.Sort(targets[k])
	}
	// 起始值固定是JFK
	result = append(result, "JFK")
	var backtracking func() bool
	backtracking = func() bool {
		// 最终结果等于机票数+1时
		if len(tickets)+1 == len(result) {
			return true
		}
		// 取出起飞航班对应的目的地
		for _, pair := range targets[result[len(result)-1]] {
			// 没有访问过的添加到结果
			if !pair.visited {
				result = append(result, pair.target)
				pair.visited = true
				// 如果能搜索到路径，返回true
				if backtracking() {
					return true
				}
				// 回溯，还原现场
				result = result[:len(result)-1]
				pair.visited = false
			}
		}
		return false
	}

	backtracking()

	return result
}
